chore[ui]: update asset names + colors & font-size
Changed the assets and colors to a darker orange to make the orange more
WCAG accessible.

Changed the font-sizes to be larger in globals.css

Changed the blues to a different shade of blue to match the orange.

Removed the width and height on the svg files themselves.

Renamed some redundancies in file names.