refactor: refactor image modals and its styles
Refactored the image modals on each page. Created a UseImageModal hook
to contain the usestates and functions for opening and closing. There is
now an imageData on every subpage.

Removed the styling for the modals on each of the pages' css.

Added a Slider component for the sliders on the pages.

Added an ImageModal component for the clicking action.